Refactored static methods in model package into a helper class
