+++ src/main/java/edu/brandeis/cs/nlp/mae/model/ExtentTag.java
@@ -28,7 +28,6 @@ import com.j256.ormlite.dao.ForeignCollection;
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.field.ForeignCollectionField;
 import com.j256.ormlite.table.DatabaseTable;
-import edu.brandeis.cs.nlp.mae.MaeStrings;
 
 import java.util.*;
 
@@ -89,7 +88,7 @@ public class ExtentTag extends Tag {
     }
 
     public List<CharIndex> setSpans(String spansString) {
-        return this.setSpans(parseSpansString(spansString));
+        return this.setSpans(ModelHelpers.parseSpansString(spansString));
     }
 
     @Override
@@ -111,7 +110,7 @@ public class ExtentTag extends Tag {
     }
 
     public String getSpansAsString() {
-        return spansToString(parseCharIndices(this.getSpansAsList()));
+        return ModelHelpers.spansToString(ModelHelpers.parseCharIndices(this.getSpansAsList()));
     }
 
     public String getText() {
@@ -122,96 +121,8 @@ public class ExtentTag extends Tag {
         this.text = text;
     }
 
-    /**
-     * Takes a string representing possibly multiple spans of an extent tag Return
-     * array of integer pairs
-     *
-     * @param spansString - string of spans
-     * @return a ArrayList of int[]
-     */
-    public static ArrayList<int[]> parseSpansString(String spansString) {
-        ArrayList<int[]> spans = new ArrayList<>();
-        if (spansString == null || spansString.equals("") || spansString.equals("-1~-1")) {
-            spans.add(new int[]{-1, -1});
-            return spans;
-        }
-
-        // split each span
-        String[] pairs = spansString.split(MaeStrings.SPANSEPARATOR);
-        for (String pair : pairs) {
-            int[] span = new int[2];
-
-            // parse start and end points
-            span[0] = Integer.parseInt(pair.split(MaeStrings.SPANDELIMITER)[0]);
-            span[1] = Integer.parseInt(pair.split(MaeStrings.SPANDELIMITER)[1]);
-
-            spans.add(span);
-        }
-        return spans;
-    }
-
-    /**
-     * Takes an array of integer pairs, then merge it into a string. Each span
-     * separated by SPANSEPARATOR start and end point of each span joined with
-     * SPANDELIMITER
-     *
-     * @param spans - an sorted set of integer pairs
-     * @return a formatted string of spans of a tag
-     */
-    public static String spansToString(ArrayList<int[]> spans) {
-        String spanString = "";
-        Iterator<int[]> iter = spans.iterator();
-        while (iter.hasNext()) {
-            int[] span = iter.next();
-            if (iter.hasNext()) {
-                spanString += span[0] + MaeStrings.SPANDELIMITER + span[1]
-                        + MaeStrings.SPANSEPARATOR;
-            } else {
-                spanString += span[0] + MaeStrings.SPANDELIMITER + span[1];
-            }
-        }
-        return spanString;
-    }
-
-    /**
-     * Takes an array of CharIndex, make it into an array of int pairs,
-     * which can be used in spansToString()
-     *
-     * @param spans - an sorted set of integer pairs
-     * @return a ArrayList of int[]
-     */
-    public static ArrayList<int[]> parseCharIndices(List<CharIndex> spans) {
-        ArrayList<int[]> spansList = new ArrayList<>();
-        if (spans == null || spans.size() ==0) {
-            spansList.add(new int[]{-1, -1});
-            return spansList;
-        }
-
-        int[] locations = new int[spans.size()];
-        for (int i = 0; i < spans.size(); i++) {
-            locations[i] = (spans.get(i).getLocation());
-        }
-        Arrays.sort(locations);
-
-        int start = locations[0];
-        int prev = locations[0];
-        for (int i = 1; i < locations.length; i++) {
-            if (i == locations.length - 1) {
-                spansList.add(new int[]{start, locations[i] + 1});
-            } else if (prev + 1 < locations[i]) {
-                spansList.add(new int[]{start, prev + 1});
-                prev = locations[i];
-                start = locations[i];
-            } else {
-                prev = locations[i];
-            }
-        }
-        return spansList;
-
-    }
-
     public Map<String, String> getAttbutesWithNames() {
-        return Tag.getAttbutesWithNames(this.getAttributes());
+        return ModelHelpers.getAttbutesWithNames(this.getAttributes());
     }
 
 }
